refactor(voice): split catalog and tests from facade
